+++ packages/web-vue/lib/color-picker-panel/type.ts
@@ -1,0 +1,13 @@
+import type { ColorFormat, HsvaColor } from '../share/type'
+
+export interface ColorPickerPanelProps {
+	current?: HsvaColor
+	formatted?: string
+	format?: ColorFormat
+	includeAlpha?: boolean
+	preset?: string[]
+}
+
+export type ColorPickerPanelEmits = {
+	change: [color: HsvaColor, event: Event]
+}