+++ packages/web-vue/lib/switch/index.vue
@@ -231,12 +231,12 @@ const getMainColor = () => {
 	}
 	return progress.value > MID_PROGRESS
 		? props.activeColor
-			? parseColor(props.activeColor)
+			? parseColor(props.activeColor)?.color
 			: disabledComputed.value
 				? getGlobalThemeColor('primary', 2)
 				: getGlobalThemeColor('primary', 6)
 		: props.inactiveColor
-			? parseColor(props.inactiveColor)
+			? parseColor(props.inactiveColor)?.color
 			: disabledComputed.value
 				? getGlobalThemeColor('neutral', 6)
 				: getGlobalThemeColor('neutral', 8)
@@ -352,7 +352,10 @@ const refresh = () => {
 	updateSize()
 }
 
-useResizeObserver(canvasWrapperRef, refresh)
+useResizeObserver(canvasWrapperRef, refresh, () => {
+	drawPixel()
+	updateSize()
+})
 useWatchGlobalCssVal(refresh)
 useTransitionEnd(canvasWrapperRef, refresh)