+++ packages/web-vue/lib/textarea/index.vue
@@ -5,6 +5,7 @@
 		:class="{
 			[`px-textarea__${sizeComputed}`]: !!sizeComputed,
 			'px-textarea__disabled': !!disabledComputed,
+			'px-textarea__readonly': !!readonlyComputed,
 			'px-textarea__resize': !!props.resize
 		}"
 		@mousedown="wrapperMousedownHandler"
@@ -301,7 +302,7 @@ onMounted(() => {
 	})
 })
 
-useResizeObserver(wrapperRef, drawPixelDebounce)
+useResizeObserver(wrapperRef, drawPixelDebounce, drawPixel)
 
 useWatchGlobalCssVal(drawPixelDebounce)