+++ packages/document/template/guide/changelog.md
@@ -8,6 +8,7 @@
 [[[zh
 ## v0.2.0
 - 新增：
+  - 新增数据输入组件：ColorPicker、Rate。
   - 日期 / 时间选择组件：DatePicker、TimePicker、DateTimePicker。
   - AutoComplete、Select 组件新增 `dropdownProps` 属性控制选项下拉组件的 Props。
   - Divider 组件新增属性 `soft` 和 `color`。
@@ -20,10 +21,12 @@
   - 修复多语言切换不会触发响应式更新的问题。
   - 修复 Slider 组件的聚焦逻辑，现在在调节器 thumb 外按下鼠标会聚焦到最近的 thumb。
   - 滚动条和它的 hook 按需导入时，现在也会导入公共样式。
+  - 修复了输入框样式的组件，只读时，边框区域鼠标指针样式不正确的问题。
 ]]]
 [[[en
 ## v0.2.0
 - New Features:
+  - New data input components: ColorPicker, Rate.
   - Date / Time selection components: DatePicker, TimePicker, DateTimePicker.
   - AutoComplete and Select components now support `dropdownProps` to control props of the option dropdown.
   - Divider component adds `soft` and `color` properties.
@@ -36,6 +39,7 @@
   - Fixed an issue where language switching did not trigger reactive updates.
   - Fixed focus logic in the Slider component: now when pressing the mouse outside the thumb, it focuses on the nearest thumb.
   - When scrollbar and its hooks are imported on demand, the public styles are now also imported.
+  - Fixed an issue where the mouse pointer style in the border area of the input-style component was incorrect when it was read-only.
 ]]]
 
 [[[zh